# docs/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "YadroService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/change-hostname": {
      "post": {
        "operationId": "YadroService_ChangeHostName",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoChangeHostNameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoChangeHostNameRequest"
            }
          }
        ],
        "tags": [
          "YadroService"
        ]
      }
    },
    "/api/list-dns": {
      "get": {
        "operationId": "YadroService_ListDNS",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoListDNSResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "YadroService"
        ]
      }
    },
    "/api/modify-dns": {
      "post": {
        "operationId": "YadroService_ModifyDNS",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoModifyDNSResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoModifyDNSRequest"
            }
          }
        ],
        "tags": [
          "YadroService"
        ]
      }
    }
  },
  "definitions": {
    "protoChangeHostNameRequest": {
      "type": "object",
      "properties": {
        "hostname": {
          "type": "string"
        }
      }
    },
    "protoChangeHostNameResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "protoListDNSResponse": {
      "type": "object",
      "properties": {
        "dnsList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "protoModifyDNSRequest": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "ip": {
          "type": "string"
        }
      }
    },
    "protoModifyDNSResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}